split ws consume & read
